+++ src/Enums/BadgeEnumInterface.php
@@ -1,0 +1,15 @@
+<?php
+/*
+ * Copyright (c) 2024. | David Annebicque | ORéOF  - All Rights Reserved
+ * @file /Users/davidannebicque/Sites/oreof/src/Enums/BadgeEnumInterface.php
+ * @author davidannebicque
+ * @project oreof
+ * @lastUpdate 30/01/2024 09:31
+ */
+
+namespace App\Enums;
+
+interface BadgeEnumInterface {
+    public function getLibelle(): string;
+    public function getBadge(): string;
+}